async load refactor

+++ Renderer/GL/OGL4/IndexBuffer.cpp
@@ -5,7 +5,7 @@ IndexBuffer::IndexBuffer(const void* data, const size_t count) :m_count(count)
 {
 	glGenBuffers(1, &m_VBO);
 	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_VBO);
-	glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(unsigned int), data, GL_STATIC_DRAW);
+	glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(unsigned short), data, GL_STATIC_DRAW);
 }
 
 IndexBuffer::IndexBuffer(
